+++ db/seeds.rb
@@ -1,7 +1,15 @@
-# This file should contain all the record creation needed to seed the database with its default values.
-# The data can then be loaded with the rails db:seed command (or created alongside the database with db:setup).
-#
-# Examples:
-#
-#   movies = Movie.create([{ name: 'Star Wars' }, { name: 'Lord of the Rings' }])
-#   Character.create(name: 'Luke', movie: movies.first)
+# frozen_string_literal: true
+
+# seed the database with a few users, events, and attendances
+puts 'Please enter your NetID'
+net_id = $stdin.gets.chomp
+
+cur_user = FactoryBot.create(:user_with_events, username: net_id)
+user2 = FactoryBot.create(:user_with_events)
+user3 = FactoryBot.create(:user_with_events)
+
+cur_user.events.each do |event|
+  2.times do
+    FactoryBot.create(:attendance, event: event)
+  end
+end