+++ CHANGELOG.md
@@ -30,6 +30,7 @@ This file should be updated before a new release is deployed.
 * Implemented Bootstrap [#154](https://gitlab.com/yale-sdmp/cardswipr/issues/154)
 * Added confirmation dialogue when deleting events [#160](https://gitlab.com/yale-sdmp/cardswipr/issues/160)
 * Added export button for exporting Event Attendances [#127](https://gitlab.com/yale-sdmp/cardswipr/issues/127)
+* Created Seed Script for Database [#148] (https://gitlab.com/yale-sdmp/cardswipr/issues/148)
 
 ### Fixed